+++ src/multi_buffer_manager.c
@@ -71,7 +71,7 @@ void nxamf_multi_buffer_manager_init(nxamf_multi_buffer_manager_t *buf, nxamf_bu
     buf->last_deserialized_index = -1;
 }
 
-size_t nxamf_multi_buffer_manager_get_last_deserialized_index(nxamf_multi_buffer_manager_t *buf)
+int nxamf_multi_buffer_manager_get_last_deserialized_index(nxamf_multi_buffer_manager_t *buf)
 {
     assert(buf != NULL);
     return buf->last_deserialized_index;
